core tests joined
