+++ testing/ecal/core_test/src/core_test.cpp
@@ -26,19 +26,22 @@
 #include <vector>
 #include <gtest/gtest.h>
 
-// subscriber callback function
-void OnReceive(const std::string& data_, long long clock_)
-{
-  std::cout << clock_ << " : " << data_ << std::endl;
-}
+namespace {
+  // subscriber callback function
+  void OnReceive(long long clock_)
+  {
+    static long long accumulated_clock = 0;
+    accumulated_clock += clock_;
+  }
 
-// timer callback function
-std::atomic_size_t     g_callback_received{ 0 };
-std::vector<long long> g_timer_vec(100);
-void OnTimer()
-{
-  if(g_callback_received < g_timer_vec.size()) g_timer_vec[g_callback_received] = eCAL::Time::GetMicroSeconds();
-  g_callback_received += 1;
+  // timer callback function
+  std::atomic_size_t     g_callback_received{ 0 };
+  std::vector<long long> g_timer_vec(100);
+  void OnTimer()
+  {
+    if (g_callback_received < g_timer_vec.size()) g_timer_vec[g_callback_received] = eCAL::Time::GetMicroSeconds();
+    g_callback_received += 1;
+  }
 }
 
 TEST(Core, MultipleInitializeFinalize)
@@ -64,7 +67,7 @@ TEST(Core, LeakedPubSub)
 
   // create subscriber and register a callback
   eCAL::string::CSubscriber<std::string> sub("foo");
-  sub.AddReceiveCallback(std::bind(OnReceive, std::placeholders::_2, std::placeholders::_4));
+  sub.AddReceiveCallback(std::bind(OnReceive, std::placeholders::_4));
 
   // start publishing thread
   eCAL::string::CPublisher<std::string> pub("foo");
@@ -73,13 +76,12 @@ TEST(Core, LeakedPubSub)
     while (!pub_stop)
     {
       pub.Send("Hello World");
-      std::this_thread::sleep_for(std::chrono::milliseconds(100));
       // some kind of busy waiting....
-      //int y = 0;
-      //for (int i = 0; i < 100000; i++)
-      //{
-      //  y += i;
-      //}
+      int y = 0;
+      for (int i = 0; i < 100000; i++)
+      {
+        y += i;
+      }
     }
   });
 
@@ -92,6 +94,58 @@ TEST(Core, LeakedPubSub)
 
   // stop publishing thread
   pub_stop = true; pub_t.join();
+}
+
+TEST(Core, CallbackDestruction)
+{
+  // initialize eCAL API
+  EXPECT_EQ(0, eCAL::Initialize(0, nullptr, "callback destruction"));
+
+  // enable loop back communication in the same thread
+  eCAL::Util::EnableLoopback(true);
+
+  // create subscriber and register a callback
+  std::shared_ptr< eCAL::string::CSubscriber<std::string>> sub;
+
+  // start publishing thread
+  eCAL::string::CPublisher<std::string> pub("foo");
+  std::atomic<bool> pub_stop(false);
+  std::thread pub_t([&]() {
+    while (!pub_stop)
+    {
+      int y = 0;
+      pub.Send("Hello World");
+      // some kind of busy waiting....
+      for (int i = 0; i < 100000; i++)
+      {
+        y += i;
+      }
+    }
+    });
+
+  std::atomic<bool> sub_stop(false);
+  std::thread sub_t([&]() {
+    while (!sub_stop)
+    {
+      sub = std::make_shared<eCAL::string::CSubscriber<std::string>>("foo");
+      sub->AddReceiveCallback(std::bind(OnReceive, std::placeholders::_4));
+      std::this_thread::sleep_for(std::chrono::seconds(2));
+    }
+    });
+
+  // let them work together
+  std::this_thread::sleep_for(std::chrono::seconds(10));
+
+  // stop publishing thread
+  pub_stop = true;
+  pub_t.join();
+
+  sub_stop = true;
+  sub_t.join();
+
+  // finalize eCAL API
+  // without destroying any pub / sub
+  EXPECT_EQ(0, eCAL::Finalize());
 }
 
 /* excluded for now, system timer jitter too high */
